refactor: remove test_conv2d_common.py, import directly from nightly
