+++ tests/ttnn/unit_tests/operations/conv/test_conv2d_l1_16384.py
@@ -4,7 +4,7 @@
 
 import pytest
 import ttnn
-from tests.ttnn.unit_tests.operations.conv.test_conv2d_common import run_conv, torch_tensor_map, HS, WS, BS
+from tests.ttnn.nightly.unit_tests.operations.conv.test_conv2d import run_conv, torch_tensor_map, HS, WS, BS
 
 pytestmark = pytest.mark.use_module_device({"l1_small_size": 16384})
